Discovered and fixed the WEIRDEST bug ever.
In example exscn3d, on Xubuntu 10.04 + FPC 2.4.0-2 + Allegro 4.4.1.1, it just
exits the program without error message after loading the texture.  To fix it
the program MUST write something in the console (i.e. "write") few lines before
the end of the "LoadTexture" function.  Note that the string to be written in
the console might be empty(i.e. "Write ('');") but must exist: it doesn't works
if you put just "Write;".  Also if you put the Write instruction inside an IF
that doesn't run it, it doesn't work. It MUST write something.